refactor: moved tests to their own folders
