--- Lab3/ts_solution/sources/queue/Car1.json
@@ -1,1 +1,0 @@
-{"id": 1, "type": "GAS", "passengers": "ROBOTS", "isDining": true, "consumption": 27}
