--- Lab3/ts_solution/sources/queue/Car10.json
@@ -1,1 +1,0 @@
-{"id": 10, "type": "ELECTRIC", "passengers": "ROBOTS", "isDining": false, "consumption": 37}
