--- Lab3/ts_solution/sources/queue/Car11.json
@@ -1,1 +1,0 @@
-{"id": 11, "type": "GAS", "passengers": "PEOPLE", "isDining": true, "consumption": 27}
