--- Lab3/ts_solution/sources/queue/Car12.json
@@ -1,1 +1,0 @@
-{"id": 12, "type": "ELECTRIC", "passengers": "ROBOTS", "isDining": false, "consumption": 29}
