--- Lab3/ts_solution/sources/queue/Car13.json
@@ -1,1 +1,0 @@
-{"id": 13, "type": "GAS", "passengers": "PEOPLE", "isDining": false, "consumption": 11}
