--- Lab3/ts_solution/sources/queue/Car14.json
@@ -1,1 +1,0 @@
-{"id": 14, "type": "ELECTRIC", "passengers": "ROBOTS", "isDining": false, "consumption": 37}
