--- Lab3/ts_solution/sources/queue/Car15.json
@@ -1,1 +1,0 @@
-{"id": 15, "type": "ELECTRIC", "passengers": "PEOPLE", "isDining": false, "consumption": 30}
