--- Lab3/ts_solution/sources/queue/Car16.json
@@ -1,1 +1,0 @@
-{"id": 16, "type": "ELECTRIC", "passengers": "ROBOTS", "isDining": true, "consumption": 26}
