--- Lab3/ts_solution/sources/queue/Car17.json
@@ -1,1 +1,0 @@
-{"id": 17, "type": "ELECTRIC", "passengers": "ROBOTS", "isDining": false, "consumption": 12}
