--- Lab3/ts_solution/sources/queue/Car18.json
@@ -1,1 +1,0 @@
-{"id": 18, "type": "ELECTRIC", "passengers": "PEOPLE", "isDining": false, "consumption": 15}
