--- Lab3/ts_solution/sources/queue/Car19.json
@@ -1,1 +1,0 @@
-{"id": 19, "type": "ELECTRIC", "passengers": "ROBOTS", "isDining": false, "consumption": 20}
