--- Lab3/ts_solution/sources/queue/Car2.json
@@ -1,1 +1,0 @@
-{"id": 2, "type": "GAS", "passengers": "ROBOTS", "isDining": false, "consumption": 18}
