--- Lab3/ts_solution/sources/queue/Car20.json
@@ -1,1 +1,0 @@
-{"id": 20, "type": "ELECTRIC", "passengers": "ROBOTS", "isDining": true, "consumption": 11}
