--- Lab3/ts_solution/sources/queue/Car21.json
@@ -1,1 +1,0 @@
-{"id": 21, "type": "GAS", "passengers": "ROBOTS", "isDining": false, "consumption": 11}
