--- Lab3/ts_solution/sources/queue/Car22.json
@@ -1,1 +1,0 @@
-{"id": 22, "type": "ELECTRIC", "passengers": "PEOPLE", "isDining": true, "consumption": 45}
