--- Lab3/ts_solution/sources/queue/Car23.json
@@ -1,1 +1,0 @@
-{"id": 23, "type": "GAS", "passengers": "ROBOTS", "isDining": true, "consumption": 11}
