--- Lab3/ts_solution/sources/queue/Car24.json
@@ -1,1 +1,0 @@
-{"id": 24, "type": "ELECTRIC", "passengers": "ROBOTS", "isDining": false, "consumption": 28}
