--- Lab3/ts_solution/sources/queue/Car25.json
@@ -1,1 +1,0 @@
-{"id": 25, "type": "ELECTRIC", "passengers": "PEOPLE", "isDining": false, "consumption": 48}
