--- Lab3/ts_solution/sources/queue/Car26.json
@@ -1,1 +1,0 @@
-{"id": 26, "type": "GAS", "passengers": "ROBOTS", "isDining": false, "consumption": 10}
