--- Lab3/ts_solution/sources/queue/Car27.json
@@ -1,1 +1,0 @@
-{"id": 27, "type": "GAS", "passengers": "PEOPLE", "isDining": false, "consumption": 26}
