--- Lab3/ts_solution/sources/queue/Car28.json
@@ -1,1 +1,0 @@
-{"id": 28, "type": "GAS", "passengers": "PEOPLE", "isDining": false, "consumption": 33}
