--- Lab3/ts_solution/sources/queue/Car29.json
@@ -1,1 +1,0 @@
-{"id": 29, "type": "GAS", "passengers": "PEOPLE", "isDining": false, "consumption": 43}
