--- Lab3/ts_solution/sources/queue/Car3.json
@@ -1,1 +1,0 @@
-{"id": 3, "type": "GAS", "passengers": "ROBOTS", "isDining": true, "consumption": 34}
