--- Lab3/ts_solution/sources/queue/Car30.json
@@ -1,1 +1,0 @@
-{"id": 30, "type": "ELECTRIC", "passengers": "PEOPLE", "isDining": false, "consumption": 50}
