--- Lab3/ts_solution/sources/queue/Car4.json
@@ -1,1 +1,0 @@
-{"id": 4, "type": "GAS", "passengers": "PEOPLE", "isDining": false, "consumption": 36}
