--- Lab3/ts_solution/sources/queue/Car5.json
@@ -1,1 +1,0 @@
-{"id": 5, "type": "GAS", "passengers": "PEOPLE", "isDining": true, "consumption": 14}
