--- Lab3/ts_solution/sources/queue/Car6.json
@@ -1,1 +1,0 @@
-{"id": 6, "type": "GAS", "passengers": "PEOPLE", "isDining": true, "consumption": 13}
