--- Lab3/ts_solution/sources/queue/Car7.json
@@ -1,1 +1,0 @@
-{"id": 7, "type": "ELECTRIC", "passengers": "ROBOTS", "isDining": true, "consumption": 25}
