--- Lab3/ts_solution/sources/queue/Car8.json
@@ -1,1 +1,0 @@
-{"id": 8, "type": "GAS", "passengers": "PEOPLE", "isDining": true, "consumption": 12}
