--- Lab3/ts_solution/sources/queue/Car9.json
@@ -1,1 +1,0 @@
-{"id": 9, "type": "ELECTRIC", "passengers": "PEOPLE", "isDining": true, "consumption": 38}
