+++ Lab3/ts_solution/src/queues/tests/linkedList.test.ts
@@ -1,7 +1,7 @@
-import { ArrayQueue } from "./arrayQueue";
-describe("Array Qeue", () => {
+import { LinkedList } from "./../linkedList";
+describe("Linked List test", () => {
   test("Insertion test", () => {
-    var queue = new ArrayQueue<number>();
+    var queue = new LinkedList<number>();
     expect(queue.isEmpty()).toBe(true);
     queue.enqueue(1);
     queue.enqueue(2);
@@ -11,7 +11,7 @@ describe("Array Qeue", () => {
     queue.enqueue(5);
   });
   test("Deletion test", () => {
-    var queue = new ArrayQueue<number>();
+    var queue = new LinkedList<number>();
     expect(queue.dequeue()).toBeUndefined();
     queue.enqueue(1);
     expect(queue.dequeue()).toBe(1);
